create modules in project
