+++ index.html
@@ -32,8 +32,7 @@
     </main>
   </body>
 
-  <script src="./index.js">
+  <script type="module" src="./main.js">
     "use strict";
-    // Код писать здесь
   </script>
 </html>
